[New] Added test cases for record form form
